+++ app/Http/Controllers/PostCommentController.php
@@ -13,6 +13,13 @@ use App\Http\Resources\Comment as CommentResource;
 
 class PostCommentController extends Controller
 {
+    public function __construct()
+    {
+        //$this->middleware('auth.role:admin', ['only' => ['blockUser']]);
+       // $this->middleware('auth.role:manager', ['only' => ['destory']]);
+        $this->middleware('auth.role:manager,user,admin', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -135,8 +142,12 @@ class PostCommentController extends Controller
 
        if($komentaras->fk_Straipsnis_id == $post_id)
        {
-            $komentaras->update($request->all());
-            return response()->json($komentaras, 200);
+            if(auth()->user()->email == Komentaras::find($id)->email || auth()->user()->role == "manager" || auth()->user()->role == "admin"){
+                $komentaras->update($request->all());
+                return new CommentResource($komentaras);
+            } else {
+                return response()->json(["message" => "You don't have permission to update this comment"], 404);
+            }
        }
        else {
             return response()->json(["message" => "You don't have permission to update this comment!"], 404);
@@ -166,11 +177,17 @@ class PostCommentController extends Controller
 
         if($komentaras->fk_Straipsnis_id == $post_id)
         {
-             $komentaras->delete();
-             return response()->json($komentaras, 200);
+            if(auth()->user()->email == Komentaras::find($id)->email || auth()->user()->role == "manager" || auth()->user()->role == "admin"){
+                $komentaras->delete();
+                return new CommentResource($komentaras);
+            } else {
+                return response()->json(["message" => "You don't have permission to delete this comment"], 404);
+            }
         }
         else {
-             return response()->json(["message" => "You don't have permission to delete this post!"], 404);
+             return response()->json(["message" => "You don't have permission to delete this comment!"], 404);
         }
+
+
     }
 }
